fix(gglib-proxy): add ContentionTimeout variant and refactor error handling
- Add ContentionTimeout(String) to ModelRuntimeError (retryable, 503)
- Add contention_timeout() helper to ErrorResponse (service_unavailable type)
- Refactor handle_runtime_error() to use suggested_status_code() (DRY)
- Update retry loop to match both ModelLoading and ContentionTimeout
- Replace Internal with ContentionTimeout in manager starvation path
- Extract models.rs tests to models_tests.rs with wire-format contract test
